HtmlBridge Phase 3 ratchet: de-list JsObjects.cs (split non-element node populators)
JsObjects.cs (1286 lines) was dominated by the ~690-line element-wrapper
ToJSObject plus three non-element node JS-wrapper populators. Split the three
populators — the minimal wrapper builders for canonical character-data nodes
(PopulateCharacterDataJSObject, for DomText/DomComment), DocumentType
(PopulateDocumentTypeJSObject) and DocumentFragment
(PopulateDocumentFragmentJSObject) — into the sibling partial
DomBridge/JsObjects.NonElementNodes.cs, leaving ToJSObject (the element wrapper
and its node-kind dispatch) behind.

Pure partial-class relocation — no signature, accessibility, or logic change,
so behaviour-identical by construction. JsObjects.cs drops 1286 -> 708, under
the 750-line guard, so its OversizedFileExemptions entry is removed. With this
de-list the live exemption set holds a single entry: LayoutMetrics.cs (2343) is
now the only HtmlBridge production file still over the limit.

Broiler.HtmlBridge.Dom builds clean; HtmlBridgeArchitectureGuardTests 14/14
green (0 offenders / 0 stale exemptions); node-wrapper coverage
(CharacterDataBindingModuleTests, DocumentFragmentSentinelMigrationTests,
NodeAccessorsBindingModuleTests, DoctypeSentinelMigrationTests,
DomImplementationTests, HtmlDomInterfacesTests, 98 tests with the guard) green.
